icons moved to icons directory
