+++ app/components/Likes/Likes.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import {LikesProps} from "@/app/components/Likes/Likes.props";
 import styles from './Likes.module.css';
 import cn from "classnames";
-import Like from '../../../public/like.svg';
+import Like from '../../../public/icons/like.svg';
 
 const Likes = ({count}: LikesProps) => {
     return (
