+++ app/components/Read/Read.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {ReadProps} from "@/app/components/Read/Read.props";
-import ReadArrow from '../../../public/arrow.svg';
+import ReadArrow from '../../../public/icons/arrow.svg';
 import styles from './Read.module.css';
 import Link from "next/link";
